% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.data.frame.substances}
\alias{as.data.frame.substances}
\alias{as.list.substances}
\title{Coerce a substance vector to a data frame or a list}
\usage{
\method{as.data.frame}{substances}(x, row.names = NULL, optional = FALSE, ...)

\method{as.list}{substances}(x, ...)
}
\arguments{
\item{x}{A \link{substances} vector.}

\item{row.names, optional}{As in \code{\link[base:as.data.frame]{base::as.data.frame()}}.}

\item{...}{Passed on.}
}
\value{
A data frame with one \code{substances} column, or a list of length-1
\code{substances} vectors.
}
\description{
Coerce a substance vector to a data frame or a list
}
\examples{
x <- set_substances(c(100, 140), c("glucose", "sodium"), "mg/dL")
as.data.frame(x)
as.list(x)
}
