% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{diff.substances}
\alias{diff.substances}
\title{Lagged differences of a substance vector}
\usage{
\method{diff}{substances}(x, ...)
}
\arguments{
\item{x}{A \link{substances} vector.}

\item{...}{Passed to \code{\link[base:diff]{base::diff()}}.}
}
\value{
A \code{substances} vector one element shorter per lag.
}
\description{
Lagged differences of a substance vector
}
\examples{
diff(set_substances(c(100, 120, 90), "glucose", "mg/dL"))
}
