% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{duplicated.substances}
\alias{duplicated.substances}
\alias{anyDuplicated.substances}
\alias{unique.substances}
\title{Duplicated and unique substance vectors}
\usage{
\method{duplicated}{substances}(x, incomparables = FALSE, ...)

\method{anyDuplicated}{substances}(x, incomparables = FALSE, ...)

\method{unique}{substances}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A \link{substances} vector.}

\item{incomparables}{Passed to the data-frame method.}

\item{...}{Passed on.}
}
\value{
As the corresponding base function.
}
\description{
Two elements match only when both the value and the substance match.
}
\examples{
x <- set_substances(c(5, 5, 5), c("glucose", "glucose", "urea"), "mmol/L")
duplicated(x)
unique(x)
}
