% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Extract.substances}
\alias{Extract.substances}
\alias{[.substances}
\alias{[[.substances}
\alias{[<-.substances}
\alias{[[<-.substances}
\title{Extract or replace parts of a substance vector}
\usage{
\method{[}{substances}(x, ...)

\method{[[}{substances}(x, ...)

\method{[}{substances}(x, ...) <- value

\method{[[}{substances}(x, ...) <- value
}
\arguments{
\item{x}{A \link{substances} vector.}

\item{...}{Index arguments, as in \link[base:Extract]{base::Extract}.}

\item{value}{Replacement value.}
}
\value{
A \code{substances} vector.
}
\description{
Extract or replace parts of a substance vector
}
\examples{
x <- set_substances(c(100, 140, 5), c("glucose", "sodium", "urea"), "mg/dL")
x[2:3]
x[[1]]
x[1] <- set_substances(120, "glucose", "mg/dL")
x
}
