% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_substance}
\alias{get_substance}
\title{Retrieve a Substance from the Database}
\usage{
get_substance(name)
}
\arguments{
\item{name}{Character. The name of the substance (e.g., \code{"water"} or \code{"glucose"}).}
}
\value{
An object of class \code{"substances"} representing the requested substance.
}
\description{
This function retrieves a predefined substance object from the internal database.
}
\examples{
water <- get_substance("water")
print(water)

}
