% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_substances_db}
\alias{get_substances_db}
\title{Get Predefined Substances Database}
\usage{
get_substances_db()
}
\value{
A named list of \code{"substances"} objects, where each key is the substance name.
}
\description{
This function returns a list of predefined substances with their molecular weights.
}
\examples{
db <- get_substances_db()
print(db$water)  # Returns the predefined substance "water"

}
