% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{mass_to_moles}
\alias{mass_to_moles}
\title{Convert Mass to Moles}
\usage{
mass_to_moles(mass, substance)
}
\arguments{
\item{mass}{Numeric. The mass of the substance in grams.}

\item{substance}{An object of class \code{"substances"} representing the chemical substance.}
}
\value{
The number of moles (numeric).
}
\description{
This function converts a given mass (in grams) to the number of moles of a substance.
}
\examples{
water <- substance("water", 18.02)
mass_to_moles(36.04, water)  # Returns 2

}
