% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{moles_to_mass}
\alias{moles_to_mass}
\title{Convert Moles to Mass}
\usage{
moles_to_mass(moles, substance)
}
\arguments{
\item{moles}{Numeric. The amount of substance in moles.}

\item{substance}{An object of class \code{"substances"} representing the chemical substance.}
}
\value{
The mass in grams (numeric).
}
\description{
This function converts a given number of moles to mass (grams).
}
\examples{
glucose <- substance("glucose", 180.16)
moles_to_mass(1, glucose)  # Returns 180.16

}
