% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rep.substances}
\alias{rep.substances}
\title{Replicate a substance vector}
\usage{
\method{rep}{substances}(x, ...)
}
\arguments{
\item{x}{A \link{substances} vector.}

\item{...}{Passed to \code{\link[base:rep]{base::rep()}}.}
}
\value{
A \code{substances} vector.
}
\description{
Replicate a substance vector
}
\examples{
rep(set_substances(100, "glucose", "mg/dL"), 3)
}
