% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.R
\name{set_units.mixed_substances}
\alias{set_units.mixed_substances}
\title{Convert a mixed_substances vector to a single unit}
\usage{
\method{set_units}{mixed_substances}(x, value, ..., mode = mixed_set_units_mode())
}
\arguments{
\item{x}{A \link{mixed_substances} vector.}

\item{value}{The target unit.}

\item{...}{Unused.}

\item{mode}{As in \code{\link[units:set_units]{units::set_units()}}.}
}
\value{
A homogeneous \link{substances} vector.
}
\description{
Convert a mixed_substances vector to a single unit
}
