% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{substance_convertible}
\alias{substance_convertible}
\title{Can these units be converted for this substance?}
\usage{
substance_convertible(from, to, substance = NA_character_, system = NULL)
}
\arguments{
\item{from, to}{Units, as strings.}

\item{substance}{A substance name, synonym or identifier, or \code{NA}.}

\item{system}{A conversion system name or object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Can these units be converted for this substance?
}
\examples{
substance_convertible("mg/dL", "mmol/L", "glucose")
substance_convertible("mg/dL", "mmol/L", NA)
}
