% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{substance_default_system}
\alias{substance_default_system}
\alias{substance_set_default_system}
\title{Get or set the default conversion system}
\usage{
substance_default_system()

substance_set_default_system(name)
}
\arguments{
\item{name}{Name of a registered system.}
}
\value{
The name of the default system.
}
\description{
Get or set the default conversion system
}
