% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance.R
\name{substance_of}
\alias{substance_of}
\alias{substance_of<-}
\alias{substance_unit}
\alias{substance_system_of}
\alias{drop_substance}
\title{Accessors for substance vectors}
\usage{
substance_of(x)

substance_of(x) <- value

substance_unit(x)

substance_system_of(x)

drop_substance(x)
}
\arguments{
\item{x}{A \code{substance} vector.}

\item{value}{Replacement value.}
}
\value{
\code{substance_of()} a character vector of substance identifiers;
\code{substance_unit()} a \code{symbolic_units}; \code{drop_substance()} a
\link[units:units]{units::units} vector; \code{substance_system_of()} the system name.
}
\description{
Accessors for substance vectors
}
\examples{
x <- substance(c(100, 140), "mg/dL", c("glucose", "sodium"))
substance_of(x)
drop_substance(x)
}
