% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substances-methods.R
\name{log,substances-method}
\alias{log,substances-method}
\title{Logarithm is Not Applicable to Substances}
\usage{
\S4method{log}{substances}(x)
}
\arguments{
\item{x}{An object of class \code{"substances"}.}
}
\value{
Throws an error.
}
\description{
This method prevents taking the logarithm of a substance object, as it is not a valid operation.
}
\examples{
\dontrun{
log(substance("water", 18.02))  # This will raise an error
}

}
