% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substances-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{set_units}
\alias{drop_units}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{units}{\code{\link[units:drop_units]{drop_units()}}, \code{\link[units:set_units]{set_units()}}}
}}

